fix: 修复秒位数字和冒号布局溢出问题
digit_x/colon_x 使用 int8_t 存储，秒位坐标(170,206)和第二个冒号(162)超出范围导致溢出，改为 int16_t。

+++ src/main.cpp
@@ -40,8 +40,8 @@ ST7789_AVR lcd(TFT_DC, TFT_RST, TFT_CS);
 GFXcanvas16 *canvas = nullptr;
 
 // 布局：6个数字的X坐标 [秒个,秒十,分个,分十,时个,时十]
-static int8_t digit_x[6];
-static int8_t colon_x[2];
+static int16_t digit_x[6];
+static int16_t colon_x[2];
 
 // 翻页动画查找表 (启动时计算，运行时零浮点)
 struct StripEntry {
